Add day 4, rudimentary multithreaded solution for day4

+++ src/sample/main.rs
@@ -9,13 +9,13 @@ fn main() -> Result<()> {
     shared_main(solution1, solution2)
 }
 
-fn solution1(input: &File) -> Result<u32> {
+fn solution1(input: &File) -> Result<u64> {
     let solution = 0;
 
     Ok(solution)
 }
 
-fn solution2(input: &File) -> Result<u32> {
+fn solution2(input: &File) -> Result<u64> {
     let solution = 0;
 
     Ok(solution)
